feat : 메뉴 개발, 여러가지 개발

+++ back-end/yaedora/src/main/java/com/yaedora/Store/dto/RatingStoreDto.java
@@ -12,7 +12,8 @@ import java.util.List;
 public class RatingStoreDto {
 
     private Long id;
-    private float rating;
+
+    private double rating;
 
     private Long member_id;
